Consolidate pings/database.ts into one sync file
